Format, mypy, pydocstyle
