+++ requirements.txt
@@ -1,5 +1,5 @@
 schema-salad >= 5.0.20200220195218, < 9
 junit-xml >= 1.8
-pytest < 8
+pytest >= 7, < 8
 defusedxml
 typing-extensions
