+++ test-requirements.txt
@@ -1,2 +1,2 @@
-pytest >= 6.2, < 7.3
+pytest >= 7, < 8
 pytest-cov
